pyproject.toml
[project]
name = "vals-agent-poker-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pokerkit>=0.6.3",
    "openai>=1.0.0",
    "model-library@git+https://github.com/vals-ai/model-proxy.git@dev",
    "dotenv>=0.9.9",
    "matplotlib>=3.8.0",
    "hydra-core>=1.3.2",
    "pyyaml>=6.0.3",
    "pytest>=8.4.2",
    "trueskill>=0.4.5",
]
